recommendation-service/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pymongo==4.10.1
pandas==2.0.3
numpy==1.24.4
scikit-surprise==1.1.4
lightfm==1.17
python-dotenv==1.0.1
fuzzywuzzy==0.18.0